fix: move GQL queries into a separate file
